refactor(editor): resolve the editor's REST transport in one place
Detekt flagged two thresholds the previous commit crossed: LongMethod on
`buildPostConfiguration` (61 of a maximum 60) and LargeClass on
`GutenbergKitSettingsBuilderTest`.

Extract the root, namespace and auth header into `resolveApiTransport`.
They are derived from one another and only valid together — a WP.com root
needs the namespace to carry `sites/<id>/`, a direct-host root needs it
empty, and the auth header has to match the host being addressed. Holding
them in one `ApiTransport` names that invariant, which is precisely what
went wrong in CMM-2383, and leaves `buildPostConfiguration` assembling the
configuration rather than deriving it.

Move the API-routing tests into `GutenbergKitSettingsBuilderApiRoutingTest`
for the same reason: the routing matrix is a coherent unit, and separating
it returns the original class to its previous size.
